// src/api/back/models/DodgeListEntryDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    UserDTO,
    UserDTOFromJSON,
    UserDTOFromJSONTyped,
    UserDTOToJSON,
} from './';

/**
 * 
 * @export
 * @interface DodgeListEntryDto
 */
export interface DodgeListEntryDto {
    /**
     * 
     * @type {UserDTO}
     * @memberof DodgeListEntryDto
     */
    user: UserDTO;
    /**
     * 
     * @type {string}
     * @memberof DodgeListEntryDto
     */
    createdAt: string;
}

export function DodgeListEntryDtoFromJSON(json: any): DodgeListEntryDto {
    return DodgeListEntryDtoFromJSONTyped(json, false);
}

export function DodgeListEntryDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): DodgeListEntryDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'user': UserDTOFromJSON(json['user']),
        'createdAt': json['createdAt'],
    };
}

export function DodgeListEntryDtoToJSON(value?: DodgeListEntryDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'user': UserDTOToJSON(value.user),
        'createdAt': value.createdAt,
    };
}